part 4 done

+++ part4/blog-list/controllers/blogs.js
@@ -1,8 +1,11 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/Blog");
+const User = require("../models/User");
+const jwt = require("jsonwebtoken");
+const config = require("../utils/config");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
@@ -23,18 +26,30 @@ blogsRouter.post("/", async (request, response, next) => {
   const body = request.body;
 
   if (!body.title || !body.url || !body.author) {
-    return response.status(400).end();
+    return response.status(400).json({ error: "title, url or author missing" });
   }
 
   try {
+    const user = request.user;
+
+    if (!user) {
+      console.log("$======400");
+      return response
+        .status(400)
+        .json({ error: "userId missing or not valid" });
+    }
+
     const blog = new Blog({
       title: body.title,
       author: body.author,
       url: body.url,
       likes: body.likes,
+      user: user._id,
     });
 
     const savedBlog = await blog.save();
+    user.blogs = user.blogs.concat(savedBlog._id);
+    await user.save();
     response.status(201).json(savedBlog);
   } catch (error) {
     next(error);
@@ -43,8 +58,22 @@ blogsRouter.post("/", async (request, response, next) => {
 
 blogsRouter.delete("/:id", async (request, response, next) => {
   try {
-    await Blog.findByIdAndDelete(request.params.id);
-    response.status(204).end();
+    const user = request.user;
+
+    if (!user) {
+      return response.status(400).json({ error: "userID missing or invalid" });
+    }
+
+    const userBlogs = user.blogs;
+
+    if (userBlogs.find((blog) => blog.toString() === request.params.id)) {
+      await Blog.findByIdAndDelete(request.params.id);
+      response.status(204).end();
+    } else {
+      response.status(400).json({
+        error: "deleting a blog is attempted by an invalid user",
+      });
+    }
   } catch (error) {
     next(error);
   }
